### 2018-08-04 #852 @arfon
:wave: @stulacy - could you take a look at the JOSS reviewers https://bit.ly/joss-reviewers and suggest a few potential candidates?
