### 2018-08-12 #852 @rrrlw
I am not familiar with multi-state modeling, so I will not be able to effectively review `multistateutils` myself. However, if there is a content expert who is willing to review, I would be happy to be a second reviewer and focus on the mechanics (installation issues, code documentation, etc.). @stulacy let me know if that would be useful to you; @arfon let me know if you would prefer people familiar with multi-state modeling to review.
