### 2018-08-13 #852 @stulacy
Yes that would be fine with me. I selected a sample of reviewers who are experienced in biostatistics in the hope that at least one would be familiar with multi-state modelling.
