### 2018-08-14 #852 @tgerke
I'm happy to help with the review as well. Like @jenniferthompson, I'm not an every day user of multistate models, but our joint efforts could likely fill the contextual review need (in addition to extra sets of eyes on the package itself, of course).
