### 2018-08-14 #893 @whedon
[REVIEW]: Estimating statistics from multi-state models using simulation with multistateutil
**Submitting author:** @stulacy (<a href="http://orcid.org/0000-0002-8570-7528">Stuart Lacy</a>)
**Repository:** <a href="https://github.com/stulacy/multistateutils" target ="_blank">https://github.com/stulacy/multistateutils</a>
**Version:** v1.2.0
