### 2018-08-15 stulacy/multistateutils#1 @rrrlw
Missing LICENSE plaintext
The DESCRIPTION file for the package does list GPL (>=2) under the License section, however, the Journal of Open Source Software requires a plain-text LICENSE file with the actual contents of the license you choose (I'm guessing GPL-2 or GPL-3) within the repository.

Can check CRAN website/manual to confirm: to avoid issues when building/installing the R package, you might have to add the LICENSE file to either .Rbuildignore or .Rinstignore (once it has been added to the repo).

Part of [this JOSS review](https://github.com/openjournals/joss-reviews/issues/893).
