### 2018-08-15 stulacy/multistateutils#2 @rrrlw
Failure to build package
R CMD build fails because package `webshot` is unavailable and not listed in the package dependencies list in `multistateutils`'s DESCRIPTION file (in Imports or Suggests). Failure happens in lines 347-349. Note that after installing `webshot`, R CMD build still fails because a relevant connection cannot be opened (screenshot below of both errors in R CMD build; second call to R CMD build was after `webshot` was installed).

![multistateutils build](https://user-images.githubusercontent.com/6846515/44130476-cf85beb4-a01b-11e8-99f6-114d4067d443.png)

Edit: issue was reproduced on Ubuntu 18 and Windows 10 machines.

Part of [this JOSS review](https://github.com/openjournals/joss-reviews/issues/893).
