### 2018-08-15 stulacy/multistateutils#3 @rrrlw
testthat causes R CMD check to fail
The `--no-build-vignettes` option was used to allow R CMD build to succeed (see Issue #2 for more details on why R CMD build fails without that option). However R CMD check fails due to an ERROR with testthat (see screenshot below).

![testthat failure](https://user-images.githubusercontent.com/6846515/44130942-dca6b1e6-a01d-11e8-805c-c04eeb904a88.png)

Part of [this JOSS review](https://github.com/openjournals/joss-reviews/issues/893).
