### 2018-08-15 stulacy/multistateutils#3 @rrrlw
I am slightly unsure about why this test is failing, because your Travis CI build appears clean. Let me know what your thoughts are, @stulacy?

(We can look at this after Issues #1 and #2, which, when solved, might fix this)

Edit: initial testing was done on Ubuntu 18, but was not reproduced on Windows 10 machine. Could be an artifact of my machine.
