### 2018-08-15 #893 @rrrlw
Major suggestions have corresponding issues in associated GitHub repo:
* [Missing LICENSE plaintext file required for JOSS](https://github.com/stulacy/multistateutils/issues/1)
* [Problem with R CMD build and undeclared dependency](https://github.com/stulacy/multistateutils/issues/2)
* [Problem with R CMD check](https://github.com/stulacy/multistateutils/issues/3) - not as certain about this one, can close issue if other reviewers don't run into the same problem

Minor suggestions (change only the ones you think will help future package users):
* Had good examples for `plot_predicted_pathway` and `predict_transitions`; would be nice if there were corresponding examples in documentation for `cohort_simulation`, `length_of_stay`, and `msprep2`
* Good practice to have package versions in `Imports` and `Suggests` fields of DESCRIPTION (e.g. `dplyr (>=0.7)`) to ensure that users with older versions of dependencies don't run into issues
* `?multistateutils::des-package` and ?multistateutils::des` open template documentation, unsure if this should be filled in or omitted
* Would be nice to add `BugReports` field in DESCRIPTION (probably link to GitHub repo Issues page)
* Would be nice to add more detail in package description itself (`?multistateutils`), can probably use text from paper.md file in JOSS submission

Other notes:
* `devtools::install_github("stulacy/multistateutils")` worked fine; examples in documentation ran fine
* As best as I can tell, functions are working correctly, but I will let the reviewers with more expertise evaluate that
* Created [pull request](https://github.com/stulacy/multistateutils/pull/4) to add DOI for a reference and add JOSS files to .Rbuildignore to prevent NOTE from R CMD check
