### 2018-08-29 stulacy/multistateutils#5 @jenniferthompson
[Gist of exact code I've been using (copied from vignette)](https://gist.github.com/jenniferthompson/3985d60366be20ba0c23309271d4c143)

This works fine for me on Ubuntu 16.04.5, matching another reviewer's experience.
