### 2018-09-02 #893 @jenniferthompson
I'm partway through, but ran into a stumbling block that I filed as an [issue in the main repo](https://github.com/stulacy/multistateutils/issues/5). (Sorry, probably should have noted that here as well!) Unfortunately this laptop is the only machine I'll have access to for the next week, but I'll see if there's more I can add.
