### 2018-09-03 stulacy/multistateutils#5 @stulacy
Hmm that's odd. I'll have a quick look at that as it worked fine on the Mac check on CRAN
