### 2018-09-03 #893 @stulacy
Sorry I've been on holiday for the last few weeks. I'll have a look at that issue and the other suggestions above
