### 2018-09-03 stulacy/multistateutils#5 @jenniferthompson
Let me know if I can help with detective work! It seemed to install fine on my machine, though I just reinstalled and noticed this warning:

```
clang++ -std=gnu++11 -I"/Library/Frameworks/R.framework/Resources/include" -DNDEBUG  -I"/Library/Frameworks/R.framework/Versions/3.5/Resources/library/Rcpp/include" -I/usr/local/include   -fPIC  -Wall -g -O2 -c simulation.cpp -o simulation.o
simulation.cpp:30:35: warning: moving a temporary object prevents copy elision [-Wpessimizing-move]
            nstate.add_transition(std::move(Transition::create_transition(trans_name, dest, as<List>(this_trans["coefs"]), tcovs)));
                                  ^
simulation.cpp:30:35: note: remove std::move call here
            nstate.add_transition(std::move(Transition::create_transition(trans_name, dest, as<List>(this_trans["coefs"]), tcovs)));
                                  ^~~~~~~~~~                                                                                     ~
1 warning generated.
```
