### 2018-09-20 7a1522d
Made change addressing compiler warning
