+++ DESCRIPTION
@@ -17,5 +17,5 @@ Suggests: flexsurv, mstate, microbenchmark, rmarkdown,
     covr
 URL: https://github.com/stulacy/multistateutils
 LinkingTo: Rcpp
-RoxygenNote: 6.0.1
+RoxygenNote: 6.1.0
 VignetteBuilder: knitr
