### 2018-09-20 stulacy/multistateutils#5 @stulacy
Can you try reinstalling from the CRAN development branch and see if that fixes it? If not, try reinstalling the `data.table` package as that could be partly causing it.
