### 2018-09-20 stulacy/multistateutils#2 @stulacy
I'm not sure about which of my dependencies requires `webshot` and why it isn't being installed automatically. I could add it to Imports if that would fix that first issue.

For the "Cannot open the connection" issue, I believe this is caused by being run as a script. I've removed this call from the Vignette and added a screenshot instead. Let me know if you can install it and run these lines in an interactive session.

On a side note, what is the preferred way of checking a package with Rcpp source? I generally used "Clean and rebuild" from RStudio then `devtools::check()`, which always ran successfully. I only got the connection error when manually running `R CMD build`.
