### 2018-09-20 stulacy/multistateutils#3 @stulacy
I've been unable to replicate this on my Debian 8 machine. It does work on Travis but that's using 14.04. I tried to test it on Ubuntu on `rhub` (16.04) but I got an error *there is no package called `flexsurv`* but I cannot understand why it isn't available.

Is there another way I can test on a remote Ubuntu 18 install? I'd like to get to the bottom of this
