### 2018-09-20 9c99668
Don't rebuild treatment pathways diagram
