+++ man/des-package.Rd
@@ -1,34 +1,0 @@
-\name{des-package}
-\alias{des-package}
-\alias{des}
-\docType{package}
-\title{
-  A short title line describing what the package does
-}
-\description{
-  A more detailed description of what the package does. A length
-  of about one to five lines is recommended.
-}
-\details{
-  This section should provide a more detailed overview of how to use the
-  package, including the most important functions.
-}
-\author{
-Your Name, email optional.
-
-Maintainer: Your Name <your@email.com>
-}
-\references{
-  This optional section can contain literature or other references for
-  background information.
-}
-\keyword{ package }
-\seealso{
-  Optional links to other man pages
-}
-\examples{
-  \dontrun{
-     ## Optional simple examples of the most important functions
-     ## These can be in \dontrun{} and \donttest{} blocks.   
-  }
-}
