### 2018-09-21 5108cf6
Added examples to msprep2 and cohort_simulation
