+++ man/cohort_simulation.Rd
@@ -57,3 +57,24 @@ A data frame with state entry times for each individual.
 \description{
 Runs a cohort discrete event simulation
 }
+\examples{
+library(multistateutils)
+library(mstate)
+library(flexsurv)
+
+# Convert data to long
+data(ebmt3)
+tmat <- trans.illdeath()
+long <- msprep(time=c(NA, 'prtime', 'rfstime'),
+               status=c(NA, 'prstat', 'rfsstat'),
+               data=ebmt3,
+               trans=tmat,
+               keep=c('age', 'dissub'))
+
+# Fit parametric models
+models <- lapply(1:3, function(i) {
+    flexsurvreg(Surv(time, status) ~ age + dissub, data=long, dist='weibull')
+})
+
+sim <- cohort_simulation(models, ebmt3, tmat)
+}
