### 2018-09-21 #893 @stulacy
I've made the following changes and pushed to the `development` branch.

> Major suggestions have corresponding issues in associated GitHub repo:
>
>     * [Missing LICENSE plaintext file required for JOSS](https://github.com/stulacy/multistateutils/issues/1)
Added LICENCE.md at directory top level

>     * [Problem with R CMD build and undeclared dependency](https://github.com/stulacy/multistateutils/issues/2)
Work in progress...

>     * [Problem with R CMD check](https://github.com/stulacy/multistateutils/issues/3) - not as certain about this one, can close issue if other reviewers don't run into the same problem
Work in progress...
>
> Minor suggestions (change only the ones you think will help future package users):
>
>     * Had good examples for `plot_predicted_pathway` and `predict_transitions`; would be nice if there were corresponding examples in documentation for `cohort_simulation`, `length_of_stay`, and `msprep2`
Added examples for these functions.

>     * Good practice to have package versions in `Imports` and `Suggests` fields of DESCRIPTION (e.g. `dplyr (>=0.7)`) to ensure that users with older versions of dependencies don't run into issues
Haven't added yet although am open to doing so if it would be beneficial for users

>     * `?multistateutils::des-package` and ?multistateutils::des` open template documentation, unsure if this should be filled in or omitted
Removed these old leftover docs.

>     * Would be nice to add `BugReports` field in DESCRIPTION (probably link to GitHub repo Issues page)
Added.

>     * Would be nice to add more detail in package description itself (`?multistateutils`), can probably use text from paper.md file in JOSS submission
Decided not to add, would prefer users to read README or vignette.
