### 2018-09-22 stulacy/multistateutils#3 @rrrlw
I think `rhub` might only be installing packages in the Imports field of DESCRIPTION and `flexsurv` is in the Suggests field. I ran R CMD check multiple times on multiple machines and the above test (predict_transitions function) only failed on a subset of the runs - is there randomization involved in that test that could be causing it to fail? If so, using `set.seed` to ensure reproducibility might be a good fix (although I would still be worried about why a test is failing in certain conditions).

A quick note: the logical AND operator is only a single ampersand (`&`) in R instead of two (`&&`) as in the failing `testthat` statement pictured above. Could that be a source of error?
