### 2018-09-22 stulacy/multistateutils#2 @rrrlw
Great! R CMD build now works perfectly for me, appears that the "Cannot open connection" issue is fixed. However, it seems that [CRAN check](https://www.r-project.org/nosvn/R.check/r-patched-solaris-x86/multistateutils-00check.html) is also having a problem with the `webshot` dependency. I think adding it to the Imports or Suggests field is reasonable (although it would likely also be a good idea to find out where it's being used).
