### 2018-09-22 stulacy/multistateutils#2 @rrrlw
Since CRAN seems to use R CMD build and R CMD check, those are what I prefer. However, I would expect that running `devtools::check()` with a clean environment would have identical behavior. Unsure what's happening...
