### 2018-09-22 #893 @rrrlw
Changes look great! Checked the corresponding boxes in this thread and responded to a couple things in repository issues page. Will recompile to ensure that updates to paper.bib made it properly to the article's PDF form, as compiled by @whedon.
