### 2018-09-22 stulacy/multistateutils#1 @rrrlw
The GPL-3 license has been added, but it has not been filled out. See the bottom section if the plaintext license file and the [GNU website](https://www.gnu.org/licenses/gpl-howto.en.html) under the section titled "The copyright notice" for more details.
