### 2018-09-24 stulacy/multistateutils#1 @stulacy
I've filled it out now, let me know if there's any other aspects that need addressing.
