### 2018-09-24 2a1c890
Corrected tests
