+++ tests/testthat/test_lengthofstay.R
@@ -7,80 +7,80 @@ models_fn <- tempfile()
 setup({
     library(mstate)
     library(flexsurv)
-    
+
     data(ebmt3)
     tmat <- trans.illdeath()
-    long <- msprep(time=c(NA, 'prtime', 'rfstime'), 
-                   status=c(NA, 'prstat', 'rfsstat'), 
-                   data=ebmt3, 
-                   trans=tmat, 
+    long <- msprep(time=c(NA, 'prtime', 'rfstime'),
+                   status=c(NA, 'prstat', 'rfsstat'),
+                   data=ebmt3,
+                   trans=tmat,
                    keep=c('age', 'dissub'))
     models <- lapply(1:3, function(i) {
         flexsurvreg(Surv(time, status) ~ age + dissub, data=long, dist='weibull')
     })
     newdata <- data.frame(age="20-40", dissub="AML")
-    
+
     saveRDS(models, models_fn)
 })
 
 teardown({
     unlink(models_fn)
-    
+
 })
 
 test_that("length_of_stay is consistent with totlos.simfs", {
     models <- readRDS(models_fn)
     newdata <- data.frame(age="20-40", dissub="AML")
     tmat <- trans.illdeath()
-    
+
     los <- length_of_stay(models, newdata, tmat, 365, start_state=1)
     simfs <- totlos.simfs(models, tmat, t=365, start=1, newdata=newdata)
-    diff <- unname(as.matrix(los[1, 5:7] / simfs))
-    expect_equal(all(diff > 0.99 && diff < 1.01), TRUE)
-    
+    diff <- unname(as.matrix(los[1, 5:7] - simfs))
+    expect_equal(all(diff > -1 & diff < 1), TRUE)
+
     los <- length_of_stay(models, newdata, tmat, 365*2, start_state=2)
     simfs <- totlos.simfs(models, tmat, t=365*2, start=2, newdata=newdata)
-    diff <- unname(as.matrix(los[1, 6:7] / simfs[2:3]))
-    expect_equal(all(diff > 0.99 && diff < 1.01), TRUE)
+    diff <- unname(as.matrix(los[1, 6:7] - simfs[2:3]))
+    expect_equal(all(diff > -1 & diff < 1), TRUE)
 })
 
 test_that("length_of_stay guards work", {
     models <- readRDS(models_fn)
     newdata <- data.frame(age="20-40", dissub="AML")
     tmat <- trans.illdeath()
-    
+
     # Models not specified
     expect_error(length_of_stay(list(3, 2), newdata, tmat, 36525, start_state=1))
-    
-    # Various issues with newdata, including wrong column names, not all covariates used by models (or 
+
+    # Various issues with newdata, including wrong column names, not all covariates used by models (or
     # by agelimit) in newdata
     expect_error(length_of_stay(models, newdata[, -1], tmat, 36525, start_state=1))
     expect_error(length_of_stay(models, data.frame(disease='CML'), tmat, 36525, start_state=1))
     expect_error(length_of_stay(models, data.frame(age='20-40'), tmat, 36525, start_state=1))
-    expect_error(length_of_stay(models, data.frame(age='20-40', dissub='CML'), 
+    expect_error(length_of_stay(models, data.frame(age='20-40', dissub='CML'),
                                 tmat, 36525, start_state=1, agelimit=36525, agecol='agecont', agescale=365.25))
-    
+
     # Problems with transition matrix
     expect_error(length_of_stay(models, newdata, tmat[-1, ], 36525, start_state=1))
     expect_error(length_of_stay(models, newdata, tmat[, -1], 36525, start_state=1))
-    
+
     # Problems with times
     expect_error(length_of_stay(models, newdata, tmat, -3, start_state=1))
[…]
