### 2018-09-24 stulacy/multistateutils#3 @stulacy
I've identified 2 causes of the issue: the use of `&&` and that the test was comparing relative error to a reference estimate, rather than absolute. I've corrected these now so it is testing the intended measure and should run for you.
