### 2018-09-24 9fdfb37
Attempt to remove warnings from Ubuntu 18 builds
