### 2018-09-24 stulacy/multistateutils#1 @rrrlw
I think you can delete the instructional text between "END OF TERMS AND CONDITIONS"  and "multistateutils:..." as well as the text after "Also add information on how to contact you..." (feel free to add an email/phone number to replace this text, if you wish).
