### 2018-09-24 stulacy/multistateutils#3 @rrrlw
Great! R CMD build and R CMD check look fine (although I would suggest switching webshot to Imports instead of Suggests).
