### 2018-09-24 stulacy/multistateutils#2 @rrrlw
Fix for #3 solves this issue as well.
