### 2018-09-24 stulacy/multistateutils#1 @rrrlw
Since the LICENSE plaintext has been added and my suggestion above is optional (the part you filled out gives sufficient information for people to know it's meant for multistateutils), I am closing this issue. (Note: I am not a lawyer by any means, so unsure about how copyright stuff works, but what you did should be sufficient to comply with JOSS policy.)
