### 2018-09-26 stulacy/multistateutils#5 @jenniferthompson
Installing the development branch seems to have done it! All the `multistateutils` functions in the Examples document run on my machine now. 🎉
