### 2018-09-26 #893 @arfon
> @arfon The fix @stulacy made recently means I can review on my laptop now, so I'll try to get back to it and have it done this weekend, if that works!

Sounds great. Thanks!
