### 2018-10-01 stulacy/multistateutils#6 @jenniferthompson
Vignette not building in development version
`vignette(package = "multistateutils")` didn't show any vignettes with the development version. Happy to close if this is just an artifact of using the development version, but the CRAN index page also lists this as "Vignette Title" so something may not be building correctly?

![image](https://user-images.githubusercontent.com/7783352/46308542-894f6380-c57f-11e8-9f04-8ddde995e5c3.png)
