### 2018-10-01 #893 @jenniferthompson
This package provides some highly useful extensions to existing capabilities in R for multistate models. Current capabilities are clearly highlighted in the Examples vignette, as well as the valuable additions provided by this package. The recently added examples are very helpful.

One issue I ran into is that the vignette doesn't seem to be building with the development version. I filed an [issue](https://github.com/stulacy/multistateutils/issues/6) in the main repo.

### Minor suggestions

- Agree with @rrrlw that adding more dependency info would be helpful; I often find it useful if I run into errors reproducing someone else's vignette code, for example. Might also be nice to explicitly note the dependencies on `data.table` and `dplyr`, as they're heavy dependencies that require some time to install.
- Future documentation could include vignette(s) for `cohort_simulations()`.
- Currently no guidelines for new contributors that I could find; this could help guide others in
contributing docs, etc - though I see @rrrlw checked this off, so possible I missed it?

Thank you for this helpful work!
