### 2018-10-02 #893 @rrrlw
@jenniferthompson, you are correct, on re-checking I failed to find contribution guidelines in the README or another file. Thank you for the correction!

Also removed text inappropriate from LICENSE file in this [pull request](https://github.com/stulacy/multistateutils/pull/7).
