### 2018-10-03 9fc2cb4
Updated installation instructions so that GitHub installs build vignette too
