### 2018-10-03 stulacy/multistateutils#6 @stulacy
The vignette not building correctly can be fixed by setting `build_vignettes=TRUE` in the `devtools::install_github()` call; I've added this information to the `README`.

The second issue with the vignette title not appearing was simply an oversight by me that I've also fixed in the development version. Let me know if the changes work for you!
