### 2018-10-03 7903220
Added code of conduct and contributing guidelines
