+++ CONTRIBUTING.md
@@ -1,0 +1,20 @@
+# Contributions
+
+Contributions to the development of `multistateutils` are always welcome and come in two main forms:
+
+  1. Reporting bugs and suggesting features
+  2. Actively contributing such changes directly to the code base
+  
+## Issues
+
+If you encounter a bug then it can be reported on the [GitHub issue tracker](https://github.com/stulacy/multistateutils/issues).
+Likewise, new feature suggestions can be made here.
+Before posting, however, check that the bug hasn't been fixed in the development version that can be installed directly from GitHub (see the [package README](https://github.com/stulacy/multistateutils/blob/master/README.md)).
+
+## Pull request
+
+If you would like to directly add features or fix issues yourself, then you can do so by means of a *pull request*.
+If you are not familiar with them then please read the [GitHub guide to creating a pull request](https://help.github.com/articles/creating-a-pull-request/).
+I would ask that pull requests target the `development` branch rather than `master`.
+
+If you have any other questions then you can always contact me directly at the email address listed in `DESCRIPTION`.
