### 2018-10-03 #893 @stulacy
I've incorporated the LICENSE pull request and believe have fixed the vignette issue (see corresponding issue thread). I've also added a CONTRIBUTING.md file with guidelines on how to contribute, suggestions to improve it are welcome.

What is a good approach for adding package versions to Imports? I could use the current release version, but I don't want to force people to update to the latest version of say `dplyr` if they have reasons for keeping older versions and the older versions work anyway.
