### 2018-10-04 stulacy/multistateutils#6 @jenniferthompson
Ah, sorry I didn't realize there was a `build_vignettes` argument that could be used (didn't click that far in the help files). Reinstalled and working!
