+++ NEWS.md
@@ -1,3 +1,9 @@
+# multistateutils 1.2.1
+
+  - Added more examples
+  - Cleaned up build process
+  - Added GPL-3 licence and contributing information
+  
 # multistateutils 1.2.0
 
   - Added `cohort_simulation` function to run discrete event simulation over populations
