### 2018-10-08 #893 @whedon
:tada::tada::tada: Congratulations on your paper acceptance! :tada::tada::tada:
